services/station/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.31.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
Faker==25.2.0